fix: BGE-290 merge MarketRepositoryWrite into MarketRepository to fix read/write lock race
GetOpenOrders() was reading world.MarketOrders without a lock while
MarketRepositoryWrite mutated the same List<T> under a separate lock instance,
allowing races and InvalidOperationException under concurrent load.

Fix: merge MarketRepositoryWrite into MarketRepository (same pattern as
ActionQueueRepository) so reads and writes share the same _lock. Also adds
world.ValidatePlayer call to CancelOrder for consistency.
