+++ src/BrowserGameEngine.StatefulGameServer.Test/TestGame/TestGame.cs
@@ -36,7 +36,6 @@ namespace BrowserGameEngine.StatefulGameServer.Test {
 		public BuildQueueRepository BuildQueueRepository { get; }
 		public BuildQueueRepositoryWrite BuildQueueRepositoryWrite { get; }
 		public MarketRepository MarketRepository { get; }
-		public MarketRepositoryWrite MarketRepositoryWrite { get; }
 		public TestWorldStateFactory WorldStateFactory { get; }
 		public GameTickModuleRegistry GameTickModuleRegistry { get; }
 		public GameTickEngine TickEngine { get; }
@@ -70,8 +69,7 @@ namespace BrowserGameEngine.StatefulGameServer.Test {
 			MessageRepositoryWrite = new MessageRepositoryWrite(Accessor, TimeProvider.System);
 			BuildQueueRepository = new BuildQueueRepository(Accessor);
 			BuildQueueRepositoryWrite = new BuildQueueRepositoryWrite(LoggerFactory.CreateLogger<BuildQueueRepositoryWrite>(), Accessor, BuildQueueRepository, AssetRepository, AssetRepositoryWrite, UnitRepository, UnitRepositoryWrite, ResourceRepository, GameDef);
-			MarketRepository = new MarketRepository(Accessor);
-			MarketRepositoryWrite = new MarketRepositoryWrite(Accessor, MarketRepository, ResourceRepository, ResourceRepositoryWrite);
+			MarketRepository = new MarketRepository(Accessor, ResourceRepository, ResourceRepositoryWrite);
 
 			var services = new ServiceCollection();
 			services.AddSingleton<IGameTickModule>(new ActionQueueExecutor(AssetRepositoryWrite));
